refactor: create root layout and unify testimonials components
  - Add root layout at src/app/layout.tsx to eliminate HTML duplication
  - Merge NewTestimonials and CanvaSections/Testimonials into single component with theme variants
  - Remove duplicate testimonial components and update imports
  - Improve component architecture with variant-based theming
