--- src/components/CanvaSections/Testimonials/Testimonial/index.ts
@@ -1,1 +1,0 @@
-export * from "./Testimonial";
