--- src/components/CanvaSections/Testimonials/Testimonial/testimonial.styles.module.css
@@ -1,112 +1,0 @@
-.blockQuoteWrapper {
-  grid-area: 1 / 1 / 2 / 2;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-bottom: var(--space-3xl);
-}
-
-.imageWrapper {
-  position: relative;
-  height: auto;
-  width: 100%;
-  max-width: 83.3%;
-  padding: 0 15px;
-  margin-bottom: var(--space-l);
-}
-
-.imageWrapper::after {
-  content: "";
-  position: absolute;
-  top: -50px;
-  right: -40px;
-  width: 100px;
-  height: 77px;
-  background: url("../../../../assets/svg/testimonial-doodle.svg") no-repeat;
-  background-size: contain;
-  pointer-events: none;
-}
-
-.img {
-  height: 100%;
-  width: 100%;
-  aspect-ratio: 4.2 / 5.6;
-}
-
-.blockquote {
-  max-width: 0 0 58.3%;
-  letter-spacing: 0.03em;
-  padding-top: 3.25rem;
-  background: url("../../../../assets/svg/quote.svg") no-repeat top left;
-  background-size: 50px 40px;
-}
-
-.blockquoteLead {
-  font-family: var(--font-ezra);
-  font-size: var(--typography-blockquote-lead);
-  line-height: 1.1;
-  font-weight: 700;
-  letter-spacing: -0.01em;
-  padding-bottom: 1rem;
-}
-
-.footer {
-  display: flex;
-  flex-direction: column;
-  margin-top: 1.5rem;
-  font-size: var(--typography-blockquote-footer);
-  font-style: normal;
-}
-
-.cite {
-  font-style: normal;
-  font-weight: 500;
-}
-
-@media (min-width: 768px) {
-  .blockQuoteWrapper {
-    flex-direction: row-reverse;
-  }
-
-  .imageWrapper {
-    max-width: 50%;
-    flex: 0 0 50%;
-    aspect-ratio: 7/9;
-    margin-bottom: 0;
-    padding: 0 0 0 15px;
-  }
-
-  .blockquote {
-    padding: 52px var(--space-xs) 0;
-    max-width: 0 0 50%;
-    flex: 0 0 50%;
-    letter-spacing: -0.01em;
-  }
-}
-
-@media (min-width: 992px) {
-  .blockQuoteWrapper {
-    justify-content: space-between;
-  }
-
-  .imageWrapper {
-    max-width: 33.3%;
-    flex: 0 0 33.333%;
-  }
-
-  .blockquote {
-    flex: 0 0 58.3%;
-    max-width: 58.3%;
-    padding: 40px 0 0 96px;
-    background-size: 80px 69px;
-  }
-}
-
-@media (min-width: 1400px) {
-  .imageWrapper::after {
-    top: -90px;
-    right: -90px;
-    width: 176px;
-    height: 134px;
-  }
-}
