--- src/components/CanvaSections/Testimonials/testimonials.styles.module.css
@@ -1,14 +1,0 @@
-.wrapper {
-  position: relative;
-}
-
-.contentWrapper {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding-top: 35px;
-}
-
-.gridWrapper {
-  display: grid;
-}
