--- src/components/NewTestimonials/NewTestimonial/index.ts
@@ -1,1 +1,0 @@
-export * from "./NewTestimonial";
