--- src/components/NewTestimonials/index.ts
@@ -1,2 +1,0 @@
-export * from "./NewTestimonials";
-export * from "./NewTestimonial";
