--- src/components/NewTestimonials/newTestimonials.styles.module.css
@@ -1,12 +1,0 @@
-.section {
-  position: relative;
-}
-
-.wrapper {
-  composes: section-container from "../../styles/composable.module.css";
-  align-items: center;
-}
-
-.gridWrapper {
-  display: grid;
-}
