+++ src/components/Testimonials/Testimonial/index.ts
@@ -1,0 +1,1 @@
+export * from "./Testimonial";
