+++ src/components/Testimonials/index.ts
@@ -1,2 +1,3 @@
 export * from "./Testimonials";
 export * from "./Testimonial";
+export * from "./Pagination";
