+++ src/components/Testimonials/testimonials.styles.module.css
@@ -1,0 +1,26 @@
+/* Home variant styles */
+.homeSection {
+  position: relative;
+}
+
+.homeWrapper {
+  composes: section-container from "../../styles/composable.module.css";
+  align-items: center;
+}
+
+/* Canva variant styles */
+.canvaWrapper {
+  position: relative;
+}
+
+.canvaContentWrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding-top: 35px;
+}
+
+/* Shared styles */
+.gridWrapper {
+  display: grid;
+}
